shared1/shared1/backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
pyttsx3
pydantic
python-multipart
